Add DocHandler abstract class unit test - Add unit test for class constructor - Move patch_dirs fixture to conftest.py to share it across tests.
